refactor: move OidcClaimExtractor into camunda-security-core
The same fail-open extraction policy is needed on the gRPC authentication
path, which lives in zeebe/gateway-grpc and cannot see the authentication
module. camunda-security-core is the nearest module both already depend on,
so the helper moves there instead of being duplicated.

This promotes camunda-security-core from test to compile scope in the
authentication module; it was already on its runtime classpath.
