+++ authentication/src/main/java/io/camunda/authentication/service/NoDBMembershipService.java
@@ -7,10 +7,10 @@
  */
 package io.camunda.authentication.service;
 
-import io.camunda.authentication.utils.OidcClaimExtractor;
 import io.camunda.security.core.oidc.OidcGroupsExtractor;
 import io.camunda.security.core.port.out.MembershipPort;
 import io.camunda.security.core.port.out.MembershipQuery;
+import io.camunda.security.oidc.OidcClaimExtractor;
 import io.camunda.security.spring.CamundaSecurityLibraryProperties;
 import io.camunda.spring.utils.ConditionalOnSecondaryStorageDisabled;
 import java.util.List;
