+++ security/security-core/src/main/java/io/camunda/security/oidc/OidcClaimExtractor.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.authentication.utils;
+package io.camunda.security.oidc;
 
 import java.util.Set;
 import java.util.concurrent.ConcurrentHashMap;
