+++ security/security-core/src/test/java/io/camunda/security/oidc/OidcClaimExtractorTest.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.authentication.utils;
+package io.camunda.security.oidc;
 
 import static org.assertj.core.api.Assertions.assertThat;
